Refactor: extend DomainTagResolver for messages
With the new feature, command handlers, commands will register DomainTags as
well. The DomainTagResolver will therefore be used by all kinds of messages:
events or commands. Therefore, extend by renaming event to message.
