+++ tests/Resolver/DomainMessage/DomainTags/Interface/InterfaceDomainTagsResolverTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Gember\EventSourcing\Test\Resolver\DomainEvent\DomainTags\Interface;
+namespace Gember\EventSourcing\Test\Resolver\DomainMessage\DomainTags\Interface;
 
-use Gember\EventSourcing\Resolver\DomainEvent\DomainTags\Interface\InterfaceDomainTagsResolver;
-use Gember\EventSourcing\Resolver\DomainEvent\DomainTags\UnresolvableDomainTagsException;
+use Gember\EventSourcing\Resolver\DomainMessage\DomainTags\Interface\InterfaceDomainTagsResolver;
+use Gember\EventSourcing\Resolver\DomainMessage\DomainTags\UnresolvableDomainTagsException;
 use Gember\EventSourcing\Test\TestDoubles\UseCase\TestUseCaseCreatedEvent;
 use Gember\EventSourcing\Test\TestDoubles\UseCase\TestUseCaseModifiedEvent;
 use Gember\EventSourcing\Test\TestDoubles\UseCase\TestDomainTag;
