+++ tests/Resolver/DomainMessage/DomainTags/UnresolvableDomainTagsCollectionExceptionTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Gember\EventSourcing\Test\Resolver\DomainEvent\DomainTags;
+namespace Gember\EventSourcing\Test\Resolver\DomainMessage\DomainTags;
 
-use Gember\EventSourcing\Resolver\DomainEvent\DomainTags\UnresolvableDomainTagsCollectionException;
-use Gember\EventSourcing\Resolver\DomainEvent\DomainTags\UnresolvableDomainTagsException;
+use Gember\EventSourcing\Resolver\DomainMessage\DomainTags\UnresolvableDomainTagsCollectionException;
+use Gember\EventSourcing\Resolver\DomainMessage\DomainTags\UnresolvableDomainTagsException;
 use Gember\EventSourcing\Test\TestDoubles\UseCase\TestUseCaseCreatedEvent;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
@@ -26,7 +26,7 @@ final class UnresolvableDomainTagsCollectionExceptionTest extends TestCase
         );
 
         self::assertSame(
-            'Unresolvable domainTags for event Gember\EventSourcing\Test\TestDoubles\UseCase\TestUseCaseCreatedEvent: It failed',
+            'Unresolvable domainTags for domain message (event/command) Gember\EventSourcing\Test\TestDoubles\UseCase\TestUseCaseCreatedEvent: It failed',
             $exception->getMessage(),
         );
         self::assertSame([$exception1, $exception2], $exception->getExceptions());
