+++ tests/Resolver/DomainMessage/DomainTags/UnresolvableDomainTagsExceptionTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Gember\EventSourcing\Test\Resolver\DomainEvent\DomainTags;
+namespace Gember\EventSourcing\Test\Resolver\DomainMessage\DomainTags;
 
-use Gember\EventSourcing\Resolver\DomainEvent\DomainTags\UnresolvableDomainTagsException;
+use Gember\EventSourcing\Resolver\DomainMessage\DomainTags\UnresolvableDomainTagsException;
 use Gember\EventSourcing\Test\TestDoubles\UseCase\TestUseCaseCreatedEvent;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
@@ -20,7 +20,7 @@ final class UnresolvableDomainTagsExceptionTest extends TestCase
         $exception = UnresolvableDomainTagsException::create(TestUseCaseCreatedEvent::class, 'It failed');
 
         self::assertSame(
-            'Unresolvable domainTags for event Gember\EventSourcing\Test\TestDoubles\UseCase\TestUseCaseCreatedEvent: It failed',
+            'Unresolvable domainTags for domain message (event/command) Gember\EventSourcing\Test\TestDoubles\UseCase\TestUseCaseCreatedEvent: It failed',
             $exception->getMessage(),
         );
     }
